refactor(ios): split background review tests
